fix: moved interface to types file
